Add: SearchContainer component for Nav
